refactor debug button out to its own function
